Combine core.rb and terraform.rb (#9541)
