+++ mmv1/provider/terraform/examples.rb
@@ -15,10 +15,9 @@ require 'uri'
 require 'api/object'
 require 'compile/core'
 require 'google/golang_utils'
-require 'provider/abstract_core'
 
 module Provider
-  class Terraform < Provider::AbstractCore
+  class Terraform
     # Generates configs to be shown as examples in docs and outputted as tests
     # from a shared template
     class Examples < Api::Object
