+++ mmv1/spec/hashicorp_copyright_spec.rb
@@ -12,7 +12,7 @@
 # limitations under the License.
 
 require 'spec_helper'
-require 'provider/core'
+require 'provider/terraform'
 
 describe 'Provider::Core.expected_output_folder?' do
   # Inputs for tests
@@ -21,7 +21,7 @@ describe 'Provider::Core.expected_output_folder?' do
   version_name = 'ga'
   start_time = Time.now
 
-  provider = Provider::Core.new(config, api, version_name, start_time)
+  provider = Provider::Terraform.new(config, api, version_name, start_time)
 
   # rubocop:disable Layout/LineLength
   it 'should identify `terraform-provider-google` as an expected output folder' do
